refactor: callDrawLine, drawLine, addGraphViewSubLayer 메서드 RecordVC -> GraphView 이동 및 리팩토링
